BTF3 Source Drop

+++ drivers/misc/samsung/scsc/platform_mif_9610.c
@@ -685,7 +685,7 @@ irqreturn_t platform_cfg_req_isr(int irq, void *data)
 	s32 ret = 0;
 	unsigned int ka_addr = 0x1000;
 	uint32_t *ka_patch_addr = ka_patch;
-	u32 id;
+	unsigned int id;
 
 #define CHECK(x) do { \
 	int retval = (x); \
@@ -1196,7 +1196,7 @@ static int platform_mif_reset(struct scsc_mif_abs *interface, bool reset)
 
 static void __iomem *platform_mif_map_region(unsigned long phys_addr, size_t size)
 {
-	int         i;
+	size_t      i;
 	struct page **pages;
 	void        *vmem;
